requirements.txt
fastapi
uvicorn
pandas
pydantic
groq
python-dotenv